更新图形系统中UICanva画布API功能 相关的Issue
原因（目的、解决的问题等）
更新图形系统中UICanva画布API功能
RM008 对标H5，支持实现如下属性及接口
RM009 线条样式属性及矩形方法
RM011 在画布上绘制文本（无填充）以及向画布上绘制图像
RM012 支持alpha和合成模式属性
RM013 渐变填充方法的实现

描述（做了什么，变更了什么）
更新图形系统中UICanva画布API功能
RM008 对标H5，支持实现如下属性及接口
RM009 线条样式属性及矩形方法
RM011 在画布上绘制文本（无填充）以及向画布上绘制图像
RM012 支持alpha和合成模式属性
RM013 渐变填充方法的实现

测试用例（新增、改动、可能影响的功能）
foundation/graphic/ui/test/unittest/components/ui_canvas_unit_test.cpp

+++ frameworks/common/text.cpp
@@ -237,7 +237,7 @@ void Text::Draw(BufferInfo& gfxDstBuffer,
         pos.y = TextPositionY(coords, (lineCount * lineHeight - style.lineSpace_));
     }
     OpacityType opa = DrawUtils::GetMixOpacity(opaScale, style.textOpa_);
-    for (int i = 0; i < lineCount; i++) {
+    for (uint16_t i = 0; i < lineCount; i++) {
         if (pos.y > mask.GetBottom()) {
             return;
         }
